+++ CHANGELOG.md
@@ -4,6 +4,7 @@
 
 ### Updated
 
+  - Profile - Billing contact
   - Resource-metadata - Too big, optimization to get it only
     on api documentation page
   - Sidebar - logo link is defined by config nuvla-logo-url